+++ Engine/KernelCore/OctreeNode.h
@@ -10,7 +10,7 @@ public:
 	~veOctreeNode();
     
     virtual bool removeChild(veNode *child) override;
-    virtual veNode* removeChild(unsigned int cIndex) override;
+    virtual veNode* removeChild(size_t cIndex) override;
 
 	bool isIn(const veBoundingBox & bbox);
